+++ src/Holding/ViewHolding/HoldingReceivingHistory/HoldingReceivingHistory.js
@@ -1,6 +1,7 @@
-import React from 'react';
+import React, { useState, useCallback, useMemo } from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
+import { orderBy } from 'lodash';
 
 import {
   Accordion,
@@ -10,6 +11,9 @@ import {
 } from '@folio/stripes/components';
 
 import useReceivingHistory from './useReceivingHistory';
+import { SORT_DIRECTION } from '../../../constants';
+
+const { ASCENDING, DESCENDING } = SORT_DIRECTION;
 
 const columnMapping = {
   'caption': <FormattedMessage id="ui-inventory.caption" />,
@@ -28,9 +32,33 @@ const columnFormatter = {
   'comment': i => i.comment || <NoValue />,
   'source': i => <FormattedMessage id={`ui-inventory.receivingHistory.source.${i.source || 'user'}`} />,
 };
+const sorters = {
+  'caption': ({ caption }) => caption,
+  'copyNumber': ({ copyNumber }) => copyNumber,
+  'enumeration': ({ enumeration }) => enumeration,
+  'receivedDate': ({ receivedDate }) => receivedDate,
+  'source': ({ source }) => source,
+};
 
 const HoldingReceivingHistory = ({ holding }) => {
+  const [sortedColumn, setSortedColumn] = useState('');
+  const [sortDirection, setSortDirection] = useState(ASCENDING);
   const { receivingHistory, isLoading } = useReceivingHistory(holding);
+
+  const data = useMemo(() => (
+    orderBy(receivingHistory, sorters[sortedColumn], sortDirection === ASCENDING ? 'asc' : 'desc')
+  ), [receivingHistory, sortedColumn, sortDirection]);
+
+  const onHeaderClick = useCallback((_e, { name: column }) => {
+    if (!sorters[column]) return;
+
+    const isChangeDirection = sortedColumn === column;
+
+    setSortedColumn(isChangeDirection ? sortedColumn : column);
+    setSortDirection(
+      isChangeDirection && sortDirection === ASCENDING ? DESCENDING : ASCENDING
+    );
+  }, [sortedColumn, sortDirection]);
 
   if (isLoading) {
     return (
@@ -49,10 +77,13 @@ const HoldingReceivingHistory = ({ holding }) => {
     >
       <MultiColumnList
         id="receiving-history-list"
-        contentData={receivingHistory}
+        contentData={data}
         visibleColumns={visibleColumns}
         columnMapping={columnMapping}
         formatter={columnFormatter}
+        sortDirection={sortDirection}
+        sortedColumn={sortedColumn}
+        onHeaderClick={onHeaderClick}
       />
     </Accordion>
   );