+++ src/constants.js
@@ -169,3 +169,8 @@ export const commands = Object.entries(commandsMap).map(([name, shortcut]) => ({
   label: <FormattedMessage id={`ui-inventory.shortcut.${name}`} />,
   shortcut,
 }));
+
+export const SORT_DIRECTION = {
+  ASCENDING: 'ascending',
+  DESCENDING: 'descending',
+};